+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/tasks/run_python.go
@@ -1,8 +1,9 @@
-package run_sh
+package tasks
 
 import (
 	"context"
 	"fmt"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/shared_utils"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/exec_result"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/files_artifacts_expansion"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
@@ -17,49 +18,61 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_validator"
 	"github.com/kurtosis-tech/stacktrace"
-	"github.com/sirupsen/logrus"
 	"github.com/xtgo/uuid"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
-	"reflect"
+	"os"
+	"path"
 	"strings"
-	"time"
 )
 
 const (
-	RunShBuiltinName = "run_sh"
+	RunPythonBuiltinName = "run_python"
 
-	ImageNameArgName  = "image"
-	RunArgName        = "run"
-	StoreFilesArgName = "store"
-	WaitArgName       = "wait"
-	FilesArgName      = "files"
+	PythonArgumentsArgName = "args"
+	PackagesArgName        = "packages"
 
-	DefaultImageName = "badouralix/curl-jq"
+	DefaultRunPythonImageName = "python:3.11-alpine"
 
-	runshCodeKey         = "code"
-	runshOutputKey       = "output"
-	runshFileArtifactKey = "files_artifacts"
-	newlineChar          = "\n"
+	runPythonCodeKey         = "code"
+	runPythonOutputKey       = "output"
+	runPythonFileArtifactKey = "files_artifacts"
+	spaceDelimiter           = " "
 
-	shellCommand = "/bin/sh"
+	shellWrapperCommand = "/bin/sh"
 
-	DefaultWaitTimeoutDurationStr = "180s"
-	DisableWaitTimeoutDurationStr = ""
+	pythonScriptFileName = "main.py"
+	pythonWorkspace      = "/tmp/python"
+
+	defaultTmpDir                  = ""
+	pythonScriptReadPermission     = 0644
+	enforceMaxSizeLimit            = true
+	temporaryPythonDirectoryPrefix = "run-python-*"
+
+	successfulPipRunExitCode = 0
 )
 
-var runTailCommandToPreventContainerToStopOnCreating = []string{"tail", "-f", "/dev/null"}
-
-func NewRunShService(serviceNetwork service_network.ServiceNetwork, runtimeValueStore *runtime_value_store.RuntimeValueStore) *kurtosis_plan_instruction.KurtosisPlanInstruction {
+func NewRunPythonService(serviceNetwork service_network.ServiceNetwork, runtimeValueStore *runtime_value_store.RuntimeValueStore) *kurtosis_plan_instruction.KurtosisPlanInstruction {
 	return &kurtosis_plan_instruction.KurtosisPlanInstruction{
 		KurtosisBaseBuiltin: &kurtosis_starlark_framework.KurtosisBaseBuiltin{
-			Name: RunShBuiltinName,
+			Name: RunPythonBuiltinName,
 
 			Arguments: []*builtin_argument.BuiltinArgument{
 				{
 					Name:              RunArgName,
 					IsOptional:        false,
 					ZeroValueProvider: builtin_argument.ZeroValueProvider[starlark.String],
+				},
+				{
+					Name:              PythonArgumentsArgName,
+					IsOptional:        true,
+					ZeroValueProvider: builtin_argument.ZeroValueProvider[*starlark.List],
+				},
+				// TODO figure out how this should handle arguments with spaces between them
+				{
+					Name:              PackagesArgName,
+					IsOptional:        true,
+					ZeroValueProvider: builtin_argument.ZeroValueProvider[*starlark.List],
 				},
 				{
 					Name:              ImageNameArgName,
@@ -93,9 +106,11 @@ func NewRunShService(serviceNetwork service_network.ServiceNetwork, runtimeValue
 		},
 
 		Capabilities: func() kurtosis_plan_instruction.KurtosisPlanInstructionCapabilities {
-			return &RunShCapabilities{
+			return &RunPythonCapabilities{
 				serviceNetwork:      serviceNetwork,
 				runtimeValueStore:   runtimeValueStore,
+				pythonArguments:     nil,
+				packages:            nil,
 				name:                "",
 				serviceConfig:       nil, // populated at interpretation time
 				run:                 "",  // populated at interpretation time
@@ -107,16 +122,18 @@ func NewRunShService(serviceNetwork service_network.ServiceNetwork, runtimeValue
 		},
 
 		DefaultDisplayArguments: map[string]bool{
-			RunArgName:        true,
-			ImageNameArgName:  true,
-			FilesArgName:      true,
-			StoreFilesArgName: true,
-			WaitArgName:       true,
+			RunArgName:             true,
+			PythonArgumentsArgName: true,
+			PackagesArgName:        true,
+			ImageNameArgName:       true,
+			FilesArgName:           true,
+			StoreFilesArgName:      true,
+			WaitArgName:            true,
 		},
 	}
 }
 
-type RunShCapabilities struct {
+type RunPythonCapabilities struct {
 	runtimeValueStore *runtime_value_store.RuntimeValueStore
 	serviceNetwork    service_network.ServiceNetwork
 
@@ -124,18 +141,34 @@ type RunShCapabilities struct {
 	name       string
 	run        string
 
+	pythonArguments []string
+	packages        []string
+
 	serviceConfig       *service.ServiceConfig
 	fileArtifactNames   []string
 	pathToFileArtifacts []string
 	wait                string
 }
 
-func (builtin *RunShCapabilities) Interpret(arguments *builtin_argument.ArgumentValuesSet) (starlark.Value, *startosis_errors.InterpretationError) {
-	runCommand, err := builtin_argument.ExtractArgumentValue[starlark.String](arguments, RunArgName)
+func (builtin *RunPythonCapabilities) Interpret(arguments *builtin_argument.ArgumentValuesSet) (starlark.Value, *startosis_errors.InterpretationError) {
+	pythonScript, err := builtin_argument.ExtractArgumentValue[starlark.String](arguments, RunArgName)
 	if err != nil {
 		return nil, startosis_errors.WrapWithInterpretationError(err, "Unable to extract value for '%s' argument", RunArgName)
 	}
-	builtin.run = runCommand.GoString()
+	builtin.run = pythonScript.GoString()
+
+	compressedScript, scriptCompressionInterpretationErr := getCompressedPythonScriptForUpload(builtin.run)
+	if err != nil {
+		return nil, scriptCompressionInterpretationErr
+	}
+	uniqueFilesArtifactName, err := builtin.serviceNetwork.GetUniqueNameForFileArtifact()
+	if err != nil {
+		return nil, startosis_errors.NewInterpretationError("an error occurred while generating unique artifact name for python script")
+	}
+	_, err = builtin.serviceNetwork.UploadFilesArtifact(compressedScript, uniqueFilesArtifactName)
+	if err != nil {
+		return nil, startosis_errors.WrapWithInterpretationError(err, "An error occurred while storing the python script to disk")
+	}
 
 	var image string
 	if arguments.IsSet(ImageNameArgName) {
@@ -145,7 +178,7 @@ func (builtin *RunShCapabilities) Interpret(arguments *builtin_argument.Argument
 		}
 		image = imageStarlark.GoString()
 	} else {
-		image = DefaultImageName
+		image = DefaultRunPythonImageName
 	}
 
 	var filesArtifactExpansion *files_artifacts_expansion.FilesArtifactsExpansion
@@ -159,10 +192,19 @@ func (builtin *RunShCapabilities) Interpret(arguments *builtin_argument.Argument
 			if interpretationErr != nil {
 				return nil, interpretationErr
 			}
+			filesArtifactMountDirPaths[pythonWorkspace] = uniqueFilesArtifactName
 			filesArtifactExpansion, interpretationErr = service_config.ConvertFilesArtifactsMounts(filesArtifactMountDirPaths, builtin.serviceNetwork)
 			if interpretationErr != nil {
 				return nil, interpretationErr
 			}
+		}
+	} else {
+		filesArtifactMountDirPaths := map[string]string{}
+		filesArtifactMountDirPaths[pythonWorkspace] = uniqueFilesArtifactName
+		var interpretationErr *startosis_errors.InterpretationError
+		filesArtifactExpansion, interpretationErr = service_config.ConvertFilesArtifactsMounts(filesArtifactMountDirPaths, builtin.serviceNetwork)
+		if interpretationErr != nil {
+			return nil, interpretationErr
 		}
 	}
 
@@ -233,20 +275,44 @@ func (builtin *RunShCapabilities) Interpret(arguments *builtin_argument.Argument
 		builtin.wait = waitTimeout
 	}
 
+	if arguments.IsSet(PackagesArgName) {
+		packagesValue, err := builtin_argument.ExtractArgumentValue[*starlark.List](arguments, PackagesArgName)
+		if err != nil {
+			return nil, startosis_errors.WrapWithInterpretationError(err, "error occurred while extracting packages information")
+		}
+		packagesList, sliceParsingErr := kurtosis_types.SafeCastToStringSlice(packagesValue, PackagesArgName)
+		if sliceParsingErr != nil {
+			return nil, startosis_errors.WrapWithInterpretationError(err, "error occurred while converting Starlark list of packages to a golang string slice")
+		}
+		builtin.packages = packagesList
+	}
+
+	if arguments.IsSet(PythonArgumentsArgName) {
+		argsValue, err := builtin_argument.ExtractArgumentValue[*starlark.List](arguments, PythonArgumentsArgName)
+		if err != nil {
+			return nil, startosis_errors.WrapWithInterpretationError(err, "error occurred while extracting passed argument information")
+		}
+		argsList, sliceParsingErr := kurtosis_types.SafeCastToStringSlice(argsValue, PythonArgumentsArgName)
+		if sliceParsingErr != nil {
+			return nil, startosis_errors.WrapWithInterpretationError(err, "error occurred while converting Starlark list of passed arguments to a golang string slice")
+		}
+		builtin.pythonArguments = argsList
+	}
+
 	resultUuid, err := builtin.runtimeValueStore.CreateValue()
 	if err != nil {
-		return nil, startosis_errors.NewInterpretationError("An error occurred while generating UUID for future reference for %v instruction", RunShBuiltinName)
+		return nil, startosis_errors.NewInterpretationError("An error occurred while generating UUID for future reference for %v instruction", RunPythonBuiltinName)
 	}
 	builtin.resultUuid = resultUuid
 	randomUuid := uuid.NewRandom()
 	builtin.name = fmt.Sprintf("task-%v", randomUuid.String())
 
-	runShCodeValue := fmt.Sprintf(magic_string_helper.RuntimeValueReplacementPlaceholderFormat, builtin.resultUuid, runshCodeKey)
-	runShOutputValue := fmt.Sprintf(magic_string_helper.RuntimeValueReplacementPlaceholderFormat, builtin.resultUuid, runshOutputKey)
+	runPythonCodeValue := fmt.Sprintf(magic_string_helper.RuntimeValueReplacementPlaceholderFormat, builtin.resultUuid, runPythonCodeKey)
+	runPythonOutputValue := fmt.Sprintf(magic_string_helper.RuntimeValueReplacementPlaceholderFormat, builtin.resultUuid, runPythonOutputKey)
 
 	dict := map[string]starlark.Value{}
-	dict[runshCodeKey] = starlark.String(runShCodeValue)
-	dict[runshOutputKey] = starlark.String(runShOutputValue)
+	dict[runPythonCodeKey] = starlark.String(runPythonCodeValue)
+	dict[runPythonOutputKey] = starlark.String(runPythonOutputValue)
 
 	// converting go slice to starlark list
 	artifactNamesList := &starlark.List{}
@@ -256,12 +322,12 @@ func (builtin *RunShCapabilities) Interpret(arguments *builtin_argument.Argument
 			_ = artifactNamesList.Append(starlark.String(name))
 		}
 	}
-	dict[runshFileArtifactKey] = artifactNamesList
+	dict[runPythonFileArtifactKey] = artifactNamesList
 	response := starlarkstruct.FromStringDict(starlarkstruct.Default, dict)
 	return response, nil
 }
 
-func (builtin *RunShCapabilities) Validate(_ *builtin_argument.ArgumentValuesSet, validatorEnvironment *startosis_validator.ValidatorEnvironment) *startosis_errors.ValidationError {
+func (builtin *RunPythonCapabilities) Validate(_ *builtin_argument.ArgumentValuesSet, validatorEnvironment *startosis_validator.ValidatorEnvironment) *startosis_errors.ValidationError {
 	if builtin.fileArtifactNames != nil {
 		if len(builtin.fileArtifactNames) != len(builtin.pathToFileArtifacts) {
 			return startosis_errors.NewValidationError("error occurred while validating file artifact name for each file in store array. "+
@@ -281,42 +347,52 @@ func (builtin *RunShCapabilities) Validate(_ *builtin_argument.ArgumentValuesSet
 	if builtin.serviceConfig.GetFilesArtifactsExpansion() != nil {
 		for _, artifactName := range builtin.serviceConfig.GetFilesArtifactsExpansion().ServiceDirpathsToArtifactIdentifiers {
 			if !validatorEnvironment.DoesArtifactNameExist(artifactName) {
-				return startosis_errors.NewValidationError("There was an error validating '%s' as artifact name '%s' does not exist", RunShBuiltinName, artifactName)
+				return startosis_errors.NewValidationError("There was an error validating '%s' as artifact name '%s' does not exist", RunPythonBuiltinName, artifactName)
 			}
 		}
 	}
+	// TODO(gm) validate Python Script(run)
 
 	validatorEnvironment.AppendRequiredContainerImage(builtin.serviceConfig.GetContainerImageName())
 	return nil
 }
 
-// Execute This is just v0 for run_sh task - we can later improve on it.
+// Execute This is just v0 for run_python task - we can later improve on it.
 //
+//	 These TODOs are copied from run-sh
 //		TODO: stop the container as soon as task completed.
-//	  Create an mechanism for other services to retrieve files from the task container
-//	  Make task as its own entity instead of currently shown under services
-func (builtin *RunShCapabilities) Execute(ctx context.Context, _ *builtin_argument.ArgumentValuesSet) (string, error) {
+//		Create an mechanism for other services to retrieve files from the task container
+//		Make task as its own entity instead of currently shown under services
+func (builtin *RunPythonCapabilities) Execute(ctx context.Context, _ *builtin_argument.ArgumentValuesSet) (string, error) {
 	_, err := builtin.serviceNetwork.AddService(ctx, service.ServiceName(builtin.name), builtin.serviceConfig)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "error occurred while creating a run_sh task with image: %v", builtin.serviceConfig.GetContainerImageName())
 	}
 
-	// create work directory and cd into that directory
-	commandToRun, err := getCommandToRun(builtin)
+	pipInstallationResult, err := setupRequiredPackages(ctx, builtin)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "an error occurred while installing dependencies")
+	}
+
+	if pipInstallationResult.GetExitCode() != successfulPipRunExitCode {
+		return "", stacktrace.NewError("an error occurred while installing dependencies as pip exited with code '%v' instead of '%v'. The error was:\n%v", pipInstallationResult.GetExitCode(), successfulPipRunExitCode, pipInstallationResult.GetOutput())
+	}
+
+	commandToRun, err := getPythonCommandToRun(builtin)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "error occurred while preparing the sh command to execute on the image")
 	}
-	fullCommandToRun := []string{shellCommand, "-c", commandToRun}
+	fullCommandToRun := []string{shellWrapperCommand, "-c", commandToRun}
 
 	// run the command passed in by user in the container
-	createDefaultDirectoryResult, err := executeWithWait(ctx, builtin, fullCommandToRun)
+	createDefaultDirectoryResult, err := executeWithWait(ctx, builtin.serviceNetwork, builtin.name, builtin.wait, fullCommandToRun)
 	if err != nil {
 		return "", stacktrace.Propagate(err, fmt.Sprintf("error occurred while executing one time task command: %v ", builtin.run))
 	}
 
 	result := map[string]starlark.Comparable{
-		runshOutputKey: starlark.String(createDefaultDirectoryResult.GetOutput()),
-		runshCodeKey:   starlark.MakeInt(int(createDefaultDirectoryResult.GetExitCode())),
+		runPythonOutputKey: starlark.String(createDefaultDirectoryResult.GetOutput()),
+		runPythonCodeKey:   starlark.MakeInt(int(createDefaultDirectoryResult.GetExitCode())),
 	}
 
 	builtin.runtimeValueStore.SetValue(builtin.resultUuid, result)
@@ -330,7 +406,7 @@ func (builtin *RunShCapabilities) Execute(ctx context.Context, _ *builtin_argume
 	}
 
 	if builtin.fileArtifactNames != nil && builtin.pathToFileArtifacts != nil {
-		err = copyFilesFromTask(ctx, builtin)
+		err = copyFilesFromTask(ctx, builtin.serviceNetwork, builtin.name, builtin.fileArtifactNames, builtin.pathToFileArtifacts)
 		if err != nil {
 			return "", stacktrace.Propagate(err, "error occurred while copying files from  a task")
 		}
@@ -338,103 +414,67 @@ func (builtin *RunShCapabilities) Execute(ctx context.Context, _ *builtin_argume
 	return instructionResult, err
 }
 
-func copyFilesFromTask(ctx context.Context, builtin *RunShCapabilities) error {
-	if builtin.fileArtifactNames == nil || builtin.pathToFileArtifacts == nil {
-		return nil
-	}
-
-	for index, fileArtifactPath := range builtin.pathToFileArtifacts {
-		fileArtifactName := builtin.fileArtifactNames[index]
-		_, err := builtin.serviceNetwork.CopyFilesFromService(ctx, builtin.name, fileArtifactPath, fileArtifactName)
+func setupRequiredPackages(ctx context.Context, builtin *RunPythonCapabilities) (*exec_result.ExecResult, error) {
+	var maybePackagesWithRuntimeValuesReplaced []string
+	for _, pythonPackage := range builtin.packages {
+		maybePackageWithRuntimeValueReplaced, err := magic_string_helper.ReplaceRuntimeValueInString(pythonPackage, builtin.runtimeValueStore)
 		if err != nil {
-			return stacktrace.Propagate(err, fmt.Sprintf("error occurred while copying file or directory at path: %v", fileArtifactPath))
+			return nil, stacktrace.Propagate(err, "an error occurred while replacing runtime value in a package passed to run_python")
 		}
+		maybePackagesWithRuntimeValuesReplaced = append(maybePackagesWithRuntimeValuesReplaced, maybePackageWithRuntimeValueReplaced)
 	}
-	return nil
-}
 
-// Copied some of the command from: exec_recipe.ResultMapToString
-// TODO: create a utility method that can be used by add_service(s) and run_sh method.
-func resultMapToString(resultMap map[string]starlark.Comparable) string {
-	exitCode := resultMap[runshCodeKey]
-	rawOutput := resultMap[runshOutputKey]
+	if len(maybePackagesWithRuntimeValuesReplaced) == 0 {
+		return nil, nil
+	}
 
-	outputStarlarkStr, ok := rawOutput.(starlark.String)
-	if !ok {
-		logrus.Errorf("Result of run_sh was not a string (was: '%v' of type '%s'). This is not fatal but the object might be malformed in CLI output. It is very unexpected and hides a Kurtosis internal bug. This issue should be reported", rawOutput, reflect.TypeOf(rawOutput))
-		outputStarlarkStr = starlark.String(outputStarlarkStr.String())
-	}
-	outputStr := outputStarlarkStr.GoString()
-	if outputStr == "" {
-		return fmt.Sprintf("Command returned with exit code '%v' with no output", exitCode)
-	}
-	if strings.Contains(outputStr, newlineChar) {
-		return fmt.Sprintf(`Command returned with exit code '%v' and the following output:
---------------------
-%v
---------------------`, exitCode, outputStr)
-	}
-	return fmt.Sprintf("Command returned with exit code '%v' and the following output: %v", exitCode, outputStr)
-}
+	packageInstallationSubCommand := fmt.Sprintf("pip install %v", strings.Join(maybePackagesWithRuntimeValuesReplaced, spaceDelimiter))
+	packageInstallationCommand := []string{shellWrapperCommand, "-c", packageInstallationSubCommand}
 
-func getCommandToRun(builtin *RunShCapabilities) (string, error) {
-	// replace future references to actual strings
-	maybeSubCommandWithRuntimeValues, err := magic_string_helper.ReplaceRuntimeValueInString(builtin.run, builtin.runtimeValueStore)
+	executionResult, err := builtin.serviceNetwork.RunExec(
+		ctx,
+		builtin.name,
+		packageInstallationCommand,
+	)
+
 	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred while replacing runtime values in run_sh")
+		return nil, stacktrace.Propagate(err, "An error occurred while installing required dependencies")
 	}
-	commandWithNoNewLines := strings.ReplaceAll(maybeSubCommandWithRuntimeValues, newlineChar, " ")
 
-	return commandWithNoNewLines, nil
+	return executionResult, nil
 }
 
-func executeWithWait(ctx context.Context, builtin *RunShCapabilities, commandToRun []string) (*exec_result.ExecResult, error) {
-	// Wait is set to None
-	if builtin.wait == DisableWaitTimeoutDurationStr {
-		return builtin.serviceNetwork.RunExec(ctx, builtin.name, commandToRun)
-	}
-
-	resultChan := make(chan *exec_result.ExecResult, 1)
-	errChan := make(chan error, 1)
-
-	timoutStr := builtin.wait
-
-	// we validate timeout string during the validation stage so it cannot be invalid at this stage
-	parsedTimeout, _ := time.ParseDuration(timoutStr)
-
-	timeDuration := time.After(parsedTimeout)
-	contextWithDeadline, cancelContext := context.WithTimeout(ctx, parsedTimeout)
-	defer cancelContext()
-
-	go func() {
-		executionResult, err := builtin.serviceNetwork.RunExec(contextWithDeadline, builtin.name, commandToRun)
+func getPythonCommandToRun(builtin *RunPythonCapabilities) (string, error) {
+	var maybePythonArgumentsWithRuntimeValueReplaced []string
+	for _, pythonArgument := range builtin.pythonArguments {
+		maybePythonArgumentWithRuntimeValueReplaced, err := magic_string_helper.ReplaceRuntimeValueInString(pythonArgument, builtin.runtimeValueStore)
 		if err != nil {
-			errChan <- err
-		} else {
-			resultChan <- executionResult
+			return "", stacktrace.Propagate(err, "an error occurred while replacing runtime value in a python arg to run_python")
 		}
-	}()
-
-	select {
-	case result := <-resultChan:
-		return result, nil
-	case err := <-errChan:
-		return nil, err
-	case <-timeDuration: // Timeout duration
-		return nil, stacktrace.NewError("The exec request timed out after %v seconds", parsedTimeout.Seconds())
+		maybePythonArgumentsWithRuntimeValueReplaced = append(maybePythonArgumentsWithRuntimeValueReplaced, maybePythonArgumentWithRuntimeValueReplaced)
 	}
+	argumentsAsString := strings.Join(maybePythonArgumentsWithRuntimeValueReplaced, spaceDelimiter)
+
+	pythonScriptAbsolutePath := path.Join(pythonWorkspace, pythonScriptFileName)
+	if len(argumentsAsString) > 0 {
+		return fmt.Sprintf("python %s %s", pythonScriptAbsolutePath, argumentsAsString), nil
+	}
+	return fmt.Sprintf("python %s", pythonScriptAbsolutePath), nil
 }
 
-func validatePathIsUniqueWhileCreatingFileArtifact(storeFiles []string) *startosis_errors.ValidationError {
-	if len(storeFiles) > 0 {
-		duplicates := map[string]uint16{}
-		for _, filePath := range storeFiles {
-			if duplicates[filePath] != 0 {
-				return startosis_errors.NewValidationError(
-					"error occurred while validating field: %v. The file paths in the array must be unique. Found multiple instances of %v", StoreFilesArgName, filePath)
-			}
-			duplicates[filePath] = 1
-		}
+func getCompressedPythonScriptForUpload(pythonScript string) ([]byte, *startosis_errors.InterpretationError) {
+	temporaryPythonScriptDir, err := os.MkdirTemp(defaultTmpDir, temporaryPythonDirectoryPrefix)
+	defer os.Remove(temporaryPythonScriptDir)
+	if err != nil {
+		return nil, startosis_errors.NewInterpretationError("an error occurred while creating a temporary folder to write the python script too")
 	}
-	return nil
+	pythonScriptFilePath := path.Join(temporaryPythonScriptDir, pythonScriptFileName)
+	if err = os.WriteFile(pythonScriptFilePath, []byte(pythonScript), pythonScriptReadPermission); err != nil {
+		return nil, startosis_errors.NewInterpretationError("an error occurred while writing python script to disk")
+	}
+	compressed, err := shared_utils.CompressPath(pythonScriptFilePath, enforceMaxSizeLimit)
+	if err != nil {
+		return nil, startosis_errors.NewInterpretationError("an error occurred while compressing the python script")
+	}
+	return compressed, nil
 }
